service_rag/requirements.txt
fastapi
uvicorn
requests
numpy
qdrant-client
openai
python-dotenv